Fix report naming issue #9 and quit if URLs list is empty
Currently, reports are given names only in regards to generation datetime.
Althought it's quite unlikely to happen, rushy might be called at the same
millisecond therefore causing the reports to have the same name - and one
of them lost. We fix this by adding worker_N prefix to report filename.

It will also optimize a bit how we handle an empty URL list by making rushy
explicitly report this and exit with code 0.